Rename minNumActivities to minNumActivitiesReached
