+++ __tests__/basic.js
@@ -2,7 +2,7 @@ import { By, Key } from 'selenium-webdriver';
 
 import { imageSnapshotOptions } from './constants.json';
 import directLineConnected from './setup/conditions/directLineConnected';
-import minNumActivities from './setup/conditions/minNumActivities';
+import minNumActivitiesReached from './setup/conditions/minNumActivitiesReached';
 import webChatLoaded from './setup/conditions/webChatLoaded';
 
 // selenium-webdriver API doc:
@@ -17,7 +17,7 @@ test('setup', async () => {
   const input = await driver.findElement(By.tagName('input[type="text"]'));
 
   await input.sendKeys('layout carousel', Key.RETURN);
-  await driver.wait(minNumActivities(2), 2000);
+  await driver.wait(minNumActivitiesReached(2), 2000);
 
   // TODO: [P2] Remove this sleep which wait for the image to be loaded
   await driver.sleep(1000);
