+++ __tests__/setup/conditions/minNumActivitiesReached.js
